Create PeginInformation class
This class is responsible for parsing a btc lock tx and obtaining the rsk address to where the funds should be transferred, and a BTC refund address in case the lock can not be completed. This information is obtained from peg-in instructions included in an output with OP_RETURN in the tx, or from the tx sender using BtcLockSender class.